[Resolvable] provides a pattern that supports resolving conflicts of identical/rivaling method names which can happen with a trait/mixin based approach for object/type composition.
It comes with two implementations. Firstly a pure [Resolvable] Trait that augments its delegated object by 4 resolve methods - [resolveBefore], [resolveAfter], [resolveAround] and  [resolveWithAlias].
Secondly a newly found 3rd variant of a JavaScript Trait, a "Silent Trait with Local Proxy" - here the [Resolvable-silent] Trait that also is able of all 4 above mentioned kinds of conflict resolutions.
But this kind of Trait does not augment the object it works on. Its main goal is preserving an objects context and delegation of this context to behavior that never will be exposed in public.